requirements.txt
virtualenv
Django
djangorestframework
celery
django-celery
django-celery-results
django-filter
django-solo
django-bootstrap3
jsonschema
argparse
requests
ConfigParser
Image
markdown
redis
pygame
coreapi
coreapi-cli
